# examples/group/openapi/openapi.json
{
 "openapi": "3.0.1",
 "info": {
  "title": "OpenAPI Demo",
  "description": "Demo of github.com/wzshiming/openapi",
  "contact": {
   "name": "wzshiming",
   "url": "https://github.com/wzshiming",
   "email": "wzshiming@foxmail.com"
  },
  "license": {
   "name": "MIT",
   "url": "https://github.com/wzshiming/openapi/blob/master/LICENSE"
  },
  "version": "0.0.1"
 },
 "servers": [
  {
   "url": "{scheme}{host}{port}{path}",
   "variables": {
    "host": {
     "enum": [
      "localhost"
     ],
     "default": "localhost"
    },
    "path": {
     "enum": [
      "/"
     ],
     "default": "/"
    },
    "port": {
     "enum": [
      ""
     ],
     "default": ""
    },
    "scheme": {
     "enum": [
      "http://",
      "https://"
     ],
     "default": "http://"
    }
   }
  }
 ],
 "paths": {
  "/item": {
   "get": {
    "tags": [
     "/item/"
    ],
    "description": "ListItem #route:\"GET /\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.2bef"
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Item.2fd3"
         }
        }
       }
      }
     },
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "/item/"
    ],
    "description": "CreateItem #route:\"PUT /\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.2bef"
     }
    ],
    "requestBody": {
     "$ref": "#/components/requestBodies/item.ff64"
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "integer",
         "format": "int64"
        }
       }
      }
     },
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   }
  },
  "/item/{itemID}": {
   "get": {
    "tags": [
     "/item/"
    ],
    "description": "GetItem #route:\"GET /{itemID}\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.2bef"
     },
     {
      "$ref": "#/components/parameters/itemID.c1ac"
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Item.2fd3"
        }
       }
      }
     },
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "/item/"
    ],
    "description": "UpdateItem #route:\"POST /{itemID}\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.2bef"
     },
     {
      "$ref": "#/components/parameters/itemID.c1ac"
     }
    ],
    "requestBody": {
     "$ref": "#/components/requestBodies/item.ff64"
    },
    "responses": {
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "/item/"
    ],
    "description": "DeleteItem #route:\"DELETE /{itemID}\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.2bef"
     },
     {
      "$ref": "#/components/parameters/itemID.c1ac"
     }
    ],
    "responses": {
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   }
  },
  "/user": {
   "put": {
    "tags": [
     "/user/"
    ],
    "description": "CreateUser #route:\"PUT /\"#\n",
    "requestBody": {
     "$ref": "#/components/requestBodies/user.a952"
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "integer",
         "format": "int64"
        }
       }
      }
     },
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   }
  },
  "/user/{userID}": {
   "get": {
    "tags": [
     "/user/"
    ],
    "description": "GetUser #route:\"GET /{userID}\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.dc86"
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/User.b847"
        }
       }
      }
     },
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "/user/"
    ],
    "description": "DeleteUser #route:\"DELETE /{userID}\"#\n",
    "parameters": [
     {
      "$ref": "#/components/parameters/userID.dc86"
     }
    ],
    "responses": {
     "400": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string",
         "format": "error"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "Item.2fd3": {
    "type": "object",
    "properties": {
     "ItemID": {
      "type": "integer",
      "description": "ItemID is the Item ID\n",
      "format": "int64"
     },
     "Message": {
      "type": "string",
      "description": "Name is the message\n"
     },
     "Name": {
      "type": "string",
      "description": "Name is the name\n"
     }
    },
    "description": "Item is a item\n"
   },
   "ItemServices.58aa": {
    "type": "object",
    "description": "ItemServices #path:\"/item/\"#\n"
   },
   "User.b847": {
    "type": "object",
    "properties": {
     "Name": {
      "type": "string",
      "description": "Name is the name\n"
     },
     "UserID": {
      "type": "integer",
      "description": "UserID is the User ID\n",
      "format": "int64"
     }
    },
    "description": "User is a user\n"
   },
   "UserServices.a448": {
    "type": "object",
    "description": "UserServices #path:\"/user/\"#\n"
   }
  },
  "responses": {
   "err.4a37": {
    "content": {
     "text/plain": {
      "schema": {
       "type": "string",
       "format": "error"
      }
     }
    }
   },
   "item.8fb9": {
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Item.2fd3"
      }
     }
    }
   },
   "itemID.a816": {
    "content": {
     "application/json": {
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    }
   },
   "items.156c": {
    "content": {
     "application/json": {
      "schema": {
       "type": "array",
       "items": {
        "$ref": "#/components/schemas/Item.2fd3"
       }
      }
     }
    }
   },
   "user.f53f": {
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/User.b847"
      }
     }
    }
   },
   "userID.0cac": {
    "content": {
     "application/json": {
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    }
   }
  },
  "parameters": {
   "itemID.c1ac": {
    "name": "itemID",
    "in": "path",
    "required": true,
    "schema": {
     "type": "integer",
     "format": "int64"
    }
   },
   "userID.2bef": {
    "name": "userID",
    "in": "query",
    "schema": {
     "type": "integer",
     "format": "int64"
    }
   },
   "userID.dc86": {
    "name": "userID",
    "in": "path",
    "required": true,
    "schema": {
     "type": "integer",
     "format": "int64"
    }
   }
  },
  "requestBodies": {
   "item.ff64": {
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Item.2fd3"
      }
     }
    }
   },
   "user.a952": {
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/User.b847"
      }
     }
    }
   }
  }
 },
 "tags": [
  {
   "name": "/user/",
   "description": "UserServices #path:\"/user/\"#\n"
  },
  {
   "name": "/user/",
   "description": "UserServices #path:\"/user/\"#\n"
  },
  {
   "name": "/user/",
   "description": "UserServices #path:\"/user/\"#\n"
  },
  {
   "name": "/item/",
   "description": "ItemServices #path:\"/item/\"#\n"
  },
  {
   "name": "/item/",
   "description": "ItemServices #path:\"/item/\"#\n"
  },
  {
   "name": "/item/",
   "description": "ItemServices #path:\"/item/\"#\n"
  },
  {
   "name": "/item/",
   "description": "ItemServices #path:\"/item/\"#\n"
  },
  {
   "name": "/item/",
   "description": "ItemServices #path:\"/item/\"#\n"
  }
 ]
}